moving _generate_queued_event_where_clause to assets.py
